Separated `gen/instruction.go` into multiple files
