+++ gen/instruction_manager.go
@@ -1,0 +1,8 @@
+package gen
+
+import "alon.kr/x/usm/core"
+
+type InstructionManager[InstT BaseInstruction] interface {
+	// Get the instruction definition that corresponds to the provided name.
+	GetInstructionDefinition(name string) (InstructionDefinition[InstT], core.ResultList)
+}
